requirements.txt
schedule~=1.2.2
requests~=2.32.3
supabase~=2.5.1
python-dotenv~=1.0.1